requirements.txt
python-vlc==3.0.12117
numpy==1.19.5
opencv-python==4.5.1.48
tensorflow==2.4.1
